Remove spectrum and image pvs until we figure out how to use with standard detector
