+++ src/dodal/beamlines/i05.py
@@ -2,13 +2,10 @@ from dodal.beamlines.i05_shared import devices as i05_shared_devices
 from dodal.common.beamlines.beamline_utils import set_beamline as set_utils_beamline
 from dodal.device_manager import DeviceManager
 from dodal.devices.beamlines.i05 import I05Goniometer
-from dodal.devices.beamlines.i05.enums import LensMode, PassEnergy
-from dodal.devices.beamlines.i05_shared import M4M5Mirror
+from dodal.devices.beamlines.i05_shared import LensMode, M4M5Mirror, PassEnergy
 from dodal.devices.common_mirror import XYZSwitchingMirror
-from dodal.devices.electron_analyser.base.energy_sources import EnergySource
 from dodal.devices.electron_analyser.mbs import MbsAnalyserDriverIO
 from dodal.devices.hutch_shutter import HutchShutter
-from dodal.devices.pgm import PlaneGratingMonochromator
 from dodal.devices.temperture_controller import Lakeshore336
 from dodal.log import set_beamline as set_log_beamline
 from dodal.utils import BeamlinePrefix, get_beamline_name
@@ -53,15 +50,9 @@ def sa() -> I05Goniometer:
 
 
 @devices.factory
-def energy_source(pgm: PlaneGratingMonochromator) -> EnergySource:
-    return EnergySource(pgm.energy.user_readback)
-
-
-@devices.factory
-def analyser_driver(energy_source: EnergySource) -> MbsAnalyserDriverIO:
+def analyser_driver() -> MbsAnalyserDriverIO:
     return MbsAnalyserDriverIO[LensMode, PassEnergy](
         prefix=f"{PREFIX.beamline_prefix}-EA-DET-02:CAM:",
         lens_mode_type=LensMode,
         pass_energy_type=PassEnergy,
-        # energy_source=energy_source,
     )
