+++ src/dodal/devices/beamlines/i05/__init__.py
@@ -1,4 +1,3 @@
-from .enums import LensMode, PassEnergy
 from .i05_motors import I05Goniometer
 
-__all__ = ["LensMode", "PassEnergy", "I05Goniometer"]
+__all__ = ["I05Goniometer"]
