+++ src/dodal/devices/beamlines/i05_shared/__init__.py
@@ -1,18 +1,14 @@
-from dodal.devices.beamlines.i05_shared.apple_knot_constants import (
+from .apple_knot_constants import (
     APPLE_KNOT_EXCLUSION_ZONES,
     energy_to_gap_converter,
     energy_to_phase_converter,
 )
-from dodal.devices.beamlines.i05_shared.compound_motors import PolynomCompoundMotors
-from dodal.devices.beamlines.i05_shared.enums import (
-    Grating,
-    M3MJ6Mirror,
-    M4M5Mirror,
-    Mj7j8Mirror,
-)
+from .compound_motors import PolynomCompoundMotors
+from .enums import Grating, LensMode, M3MJ6Mirror, M4M5Mirror, Mj7j8Mirror, PassEnergy
 
 __all__ = [
     "Grating",
+    "LensMode",
     "Mj7j8Mirror",
     "M3MJ6Mirror",
     "M4M5Mirror",
@@ -20,4 +16,5 @@ __all__ = [
     "energy_to_gap_converter",
     "energy_to_phase_converter",
     "APPLE_KNOT_EXCLUSION_ZONES",
+    "PassEnergy",
 ]
